Add files via upload
This is a dummy file for an app called Connection, just make a new application from android studio, Make a button and a texview. Give the button the id "bn" and give the textview the id "txt" then just copy paste this MainActivity java file as soon as you add the dependency for Android Volley.